152.乘积最大的子数组/main_test.go
package main

import "testing"

func Test_maxProduct(t *testing.T) {
	type args struct {
		nums []int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			name: "nums = [2,3,-2,4]",
			args: args{
				nums: []int{2, 3, -2, 4},
			},
			want: 6,
		},
		{
			name: "nums = [-2,0,-1]",
			args: args{
				nums: []int{-2, 0, -1},
			},
			want: 0,
		},
		{
			name: "nums = [-2, -1, -3, -4, -5,0,-1]",
			args: args{
				nums: []int{-2, -1, -3, -4, -5, 0, -1},
			},
			want: 60,
		},
		{
			name: "nums = [0,2]",
			args: args{
				nums: []int{0, 2},
			},
			want: 2,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := maxProduct(tt.args.nums); got != tt.want {
				t.Errorf("maxProduct() = %v, want %v", got, tt.want)
			}
		})
	}
}
